backend/requirements.txt
# Перед установкой: python -m pip install --upgrade pip setuptools wheel
# (иначе старый pip падает на cryptography/pyproject.toml)
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# bcrypt 5.x + passlib 1.7.4: ошибка при инициализации (72 bytes)
bcrypt==4.0.1
python-multipart==0.0.17
email-validator==2.2.0
typing_extensions>=4.12.0
